[#19] refactor: Place, SearchedPlace 리팩토링 revert
